Update: 가게 수정시 이미지 null 검증 추가

+++ src/main/java/com/example/withdogandcat/domain/shop/ShopService.java
@@ -94,7 +94,7 @@ public class ShopService {
             throw new CustomException(ErrorCode.ACCOUNT_NOT_FOUND);
         }
 
-        if (!imageFiles.isEmpty() && imageFiles.stream().anyMatch(file -> !file.isEmpty())) {
+        if (imageFiles != null && !imageFiles.isEmpty() && imageFiles.stream().anyMatch(file -> !file.isEmpty())) {
             imageS3Service.deleteImages(shop.getImages());
             shop.clearImages();
             List<Image> newImages = imageS3Service.uploadMultipleImages(imageFiles, shop);
